+++ web/src/site-admin/SiteAdminCodeIntelligencePage.tsx
@@ -1,0 +1,24 @@
+import * as React from 'react'
+import { PageTitle } from '../components/PageTitle'
+import { eventLogger } from '../tracking/eventLogger'
+import { SiteAdminLangServers } from './SiteAdminLangServers'
+
+/**
+ * A page displaying information about code intelligence on this site.
+ */
+export class SiteAdminCodeIntelligencePage extends React.PureComponent {
+    public componentDidMount(): void {
+        eventLogger.logViewEvent('SiteAdminCodeIntelligence')
+    }
+
+    public render(): JSX.Element {
+        return (
+            <div className="site-admin-detail-list">
+                <div className="site-admin-detail-list">
+                    <PageTitle title="Code Intelligence" />
+                    <SiteAdminLangServers />
+                </div>
+            </div>
+        )
+    }
+}